Implement apply fees as u64 for IOC that don't match

+++ programs/openbook-v2/tests/program_test/client.rs
@@ -225,7 +225,7 @@ pub struct CreateMarketInstruction {
     pub base_lot_size: i64,
     pub maker_fee: f32,
     pub taker_fee: f32,
-    pub fee_penalty: f32,
+    pub fee_penalty: u64,
     pub settle_fee_flat: f32,
     pub settle_fee_amount_threshold: f32,
 }
